refactor: move verb constants to new Verbs class in annotations module
